fix: repair failing tests and code_graph uint schema
- Add missing commit_title to submit_work test payload (required field)
- Fix microcompact test to account for 2 assistant messages per logical
  turn when checking exempt tool results
- Change code_graph limit param from usize to i64 for MCP schema
  compliance (no uint format)
- Update MCP tools schema snapshot

+++ server/crates/djinn-mcp/src/tools/graph_tools.rs
@@ -31,7 +31,7 @@ pub struct CodeGraphParams {
     pub kind_filter: Option<String>,
     /// Maximum results for `ranked` (default 20) or max traversal depth for `impact` (default 3).
     #[serde(default)]
-    pub limit: Option<usize>,
+    pub limit: Option<i64>,
 }
 
 // ── Response types ──────────────────────────────────────────────────────────────
@@ -157,7 +157,7 @@ impl DjinnMcpServer {
         params: &CodeGraphParams,
     ) -> Result<CodeGraphResponse, String> {
         validate_kind_filter(params.kind_filter.as_deref())?;
-        let limit = params.limit.unwrap_or(20);
+        let limit = params.limit.unwrap_or(20) as usize;
         let nodes = self
             .state
             .repo_graph()
@@ -189,7 +189,7 @@ impl DjinnMcpServer {
         params: &CodeGraphParams,
     ) -> Result<CodeGraphResponse, String> {
         let key = require_key(params)?;
-        let depth = params.limit.unwrap_or(3);
+        let depth = params.limit.unwrap_or(3) as usize;
         let impact = self
             .state
             .repo_graph()
